services/api/pyproject.toml
[project]
name = "cad-agent-api"
version = "0.2.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "sqlalchemy==2.0.34",
  "psycopg[binary]==3.2.1",
  "pydantic==2.8.2",
  "pydantic-settings==2.4.0",
  "pyyaml==6.0.2",
  "httpx==0.27.2",
  "redis==5.0.8",
  "rq==1.16.2",
  "boto3==1.34.162",
  "pgvector==0.3.2",
]
[project.optional-dependencies]
test = [
  "pytest==8.3.2",
  "pytest-asyncio==0.24.0",
]
[tool.pytest.ini_options]
testpaths = ["app/tests"]
